+++ grapesy/test-grapesy/Test/Sanity/Disconnect.hs
@@ -205,18 +205,32 @@ test_serverDisconnect = withTemporaryFile $ \ipcFile -> do
             go :: Int -> Client.ReconnectPolicy
             go n
               | n == 5
-              = Client.ReconnectAfter def $ do
+              = Client.ReconnectPolicy $ do
                   killRestarted <- startServer
                   port2 <- ipcRead
                   putMVar signalRestart killRestarted
-                  return $
-                    Client.ReconnectAfter
-                      (Client.ReconnectToNew $ serverAddress port2)
-                      (pure Client.DontReconnect)
+                  return $ Client.DoReconnect Client.Reconnect {
+                      Client.nextPolicy =
+                        Client.ReconnectPolicy $
+                          pure $ Client.DoReconnect Client.Reconnect {
+                                Client.reconnectTo =
+                                  Client.ReconnectToNew $ serverAddress port2
+                              , Client.nextPolicy =
+                                  Client.ReconnectPolicy $ pure Client.DontReconnect
+                              , Client.onReconnect = Nothing
+                              }
+                      , Client.reconnectTo = Client.ReconnectToOriginal
+                      , Client.onReconnect = def
+                    }
               | otherwise
-              = Client.ReconnectAfter def $ do
+              = Client.ReconnectPolicy $ do
                   threadDelay 10000
-                  return $ go (n + 1)
+                  return $
+                    Client.DoReconnect Client.Reconnect {
+                        reconnectTo = Client.ReconnectToOriginal
+                      , onReconnect = def
+                      , nextPolicy  = go (n + 1)
+                      }
 
         connParams :: Client.ConnParams
         connParams = def { Client.connReconnectPolicy = reconnectPolicy }
@@ -411,4 +425,3 @@ type TrivialRpc rpc = (
 
 type RPC1 = Trivial' "rpc1"
 type RPC2 = Trivial' "rpc2"
-